simplify tooltip
